new file parsing part1 check file
