+++ maptest.cub
@@ -1,0 +1,3 @@
+oui
+mais
+non
